Create Type CASTs
Type CASTs
In the video, you saw that type casts are a possible solution for data type issues. If you know that a certain column stores numbers as text, you can cast the column to a numeric form, i.e. to integer.

SELECT CAST(some_column AS integer)
FROM table;
Now, the some_column column is temporarily represented as integer instead of text, meaning that you can perform numeric calculations on the column.

Instructions

Execute the given sample code.
As it doesn't work, add an integer type cast at the right place and execute it again.