Create Disallow NULL values with SET NOT NULL 1
Disallow NULL values with SET NOT NULL
The professors table is almost ready now. However, it still allows for NULLs to be entered. Although some information might be missing about some professors, there's certainly columns that always need to be specified.

Instructions 1/2

Add a not-null constraint for the firstname column.